refactor(files): #373 Phase 4-1 commands/files.rs を 3 sub-module に分割
Issue #373 / Phase 4-1: `commands/files.rs` (790 行) を `commands/files/`
配下の 3 sub-module に責務別分割。move only / bit identical / 既存 IPC
コマンド名と関数 signature は不変。

# 構成

```
commands/files.rs              790 → 336 行 (struct 定義 + 3 Tauri command)
commands/files/
  encoding.rs        新規  255 行  text encoding 関連 (encode_for_save / detect_text_or_binary
                                  / utf16_decode / utf32_decode + 2 test mod)
  path_safety.rs     新規  181 行  safe_join + safe_join_tests (#[cfg(unix)] symlink test 含む)
  hash.rs            新規   20 行  sha256_hex + mtime_ms_of
```

# 公開 API の維持

- `safe_join` は `commands/git.rs` から `crate::commands::files::safe_join` で
  外部呼び出しされているため、`pub use path_safety::safe_join` で再 export
- 3 つの `#[tauri::command]` (`files_list` / `files_read` / `files_write`) の
  signature は **1 文字も変えない** → `lib.rs` の `invoke_handler!` 登録および
  IPC コマンド名は不変
- 隣接ファイル (`atomic_write` / `fs_watch` / `role_profiles`) は touch せず

# 検証結果

- `cargo check`: ✅ green
- `cargo test --lib`: ✅ 83 passed / 0 failed (移動した 15 test もすべて pass)
  - `commands::files::encoding::detect_tests::*` (4)
  - `commands::files::encoding::encode_tests::*` (6)
  - `commands::files::path_safety::safe_join_tests::*` (5)
- `cargo clippy --no-deps`: ✅ 13 warnings = baseline subset (新規警告 0)
  baseline #2/#3/#4 は move 先 (`files/encoding.rs` / `files/path_safety.rs`) に追従

# Agent Team で並列分担

3 sub-module の draft を 3 人の teammate (encoding / path-safety / hash-helpers
extractor) に並列で書かせ、team-lead 側で:
1. files.rs の冒頭に `mod encoding; mod hash; mod path_safety;` を追加
2. `pub use path_safety::safe_join;` で外部 API を維持
3. 関数本体を sub-module へ移動した分を削除し、Write で全面再構築
4. `cargo check` / `cargo test` / `cargo clippy` で確認

を逐次実施。

Refs #373
