+++ src-tauri/src/commands/files/hash.rs
@@ -1,0 +1,20 @@
+use sha2::{Digest, Sha256};
+
+/// Issue #119: バイト列の SHA-256 を 16 進文字列で返す。
+pub fn sha256_hex(bytes: &[u8]) -> String {
+    let mut h = Sha256::new();
+    h.update(bytes);
+    let digest = h.finalize();
+    let mut out = String::with_capacity(digest.len() * 2);
+    for b in digest.iter() {
+        out.push_str(&format!("{:02x}", b));
+    }
+    out
+}
+
+pub fn mtime_ms_of(meta: &std::fs::Metadata) -> Option<u64> {
+    meta.modified()
+        .ok()
+        .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
+        .map(|d| d.as_millis() as u64)
+}
